setup.rst: also do automatic conversion
This applies the same automatic conversion to the original setup.txt as
was applied to the other .txt files. This ensures consistency between
all the .rst files. This commit also serves to show what the automatic
conversion did wrong (for example removing the top comments and chapter
heading) so that we can go and fix those things everywhere.

On the plus side, this also shows that the automatic conversion actually
doesn't deviate that much from what kimrhh did manually.