handbook: rename master document to index.rst
ReadTheDocs assumes that the master document is called index.rst, so
follow that convention. Various more-or-less hacky workarounds (creating
a symlink, creating index.rst with just a ".. include" directive) caused
build warnings which would, in the long run, be more annoying and
confusing than having the name of the master document dictated to us.